# specs/001-starter-app-template/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Starter App Template API
  description: API specification for the Next.js starter app template with Supabase, Stripe, and RAG integrations
  version: 1.0.0
  contact:
    name: Starter Template Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-app.vercel.app/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        file_name:
          type: string
        file_size:
          type: integer
          minimum: 1
        file_type:
          type: string
          enum: [application/pdf, text/plain, text/markdown]
        upload_status:
          type: string
          enum: [uploading, uploaded, failed]
        processing_status:
          type: string
          enum: [pending, processing, completed, failed]
        chunk_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /documents:
    get:
      summary: List user documents
      description: Retrieve all documents uploaded by the authenticated user
      parameters:
        - name: status
          in: query
          description: Filter by processing status
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          description: Maximum number of documents to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of documents to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      summary: Upload document
      description: Upload a new document for processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  maxLength: 255
              required:
                - file
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid file or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /documents/{documentId}:
    get:
      summary: Get document details
      description: Retrieve details for a specific document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete document
      description: Delete a document and all associated chunks
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /stripe/checkout:
    post:
      summary: Create Stripe checkout session
      description: Create a Stripe checkout session for subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                price_id:
                  type: string
                  description: Stripe price ID
                success_url:
                  type: string
                  format: uri
                cancel_url:
                  type: string
                  format: uri
              required:
                - price_id
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                    format: uri
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stripe/portal:
    post:
      summary: Create customer portal session
      description: Create a Stripe customer portal session for subscription management
      responses:
        '200':
          description: Portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  portal_url:
                    type: string
                    format: uri
        '400':
          description: No active subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /email/subscribe:
    post:
      summary: Subscribe to email list
      description: Subscribe user to marketing email list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                  maxLength: 50
                tags:
                  type: array
                  items:
                    type: string
                    maxLength: 50
              required:
                - email
      responses:
        '200':
          description: Successfully subscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid email or subscription failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'